Refactor: move hivemind.client.optim to hivemind.optim
