+++ hivemind/__init__.py
@@ -2,5 +2,6 @@ from hivemind.client import *
 from hivemind.dht import *
 from hivemind.server import *
 from hivemind.utils import *
+from hivemind.optim import *
 
 __version__ = '0.9.7'
