+++ hivemind/client/__init__.py
@@ -2,4 +2,3 @@ from hivemind.client.expert import RemoteExpert
 from hivemind.client.moe import RemoteMixtureOfExperts
 from hivemind.client.averaging import DecentralizedAverager
 from hivemind.client.averaging.training import TrainingAverager
-from hivemind.client.optim import ParameterAveragingOptimizer, DecentralizedSGD, CollaborativeOptimizer
