--- hivemind/client/optim/__init__.py
@@ -1,2 +1,0 @@
-from hivemind.client.optim.simple import ParameterAveragingOptimizer, DecentralizedSGD
-from hivemind.client.optim.collaborative import CollaborativeOptimizer
