+++ hivemind/optim/simple.py
@@ -6,7 +6,7 @@ import torch
 
 from hivemind.dht import DHT
 from hivemind.client.averaging import DecentralizedAverager
-from hivemind.client.optim.base import DecentralizedOptimizerBase
+from hivemind.optim.base import DecentralizedOptimizerBase
 from hivemind.utils import get_logger, get_dht_time
 
 logger = get_logger(__name__)
